+++ Heartland-iOS-SDK/Classes/TapToPay/TapToPay.swift
@@ -1,0 +1,17 @@
+//
+//
+//  TapToPay.swift
+//  Heartland-iOS-SDK
+//
+    
+
+import Foundation
+import SwiftUI
+
+public class TapToPay {
+    public init() { }
+    
+    public func tapToPayViewController() -> UIViewController {
+        UIHostingController(rootView: TapToPayView(model: TapToPayViewModel()))
+    }
+}